% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colMeansRcpp}
\alias{colMeansRcpp}
\title{Function to calculate column means of a NumericMatrix}
\usage{
colMeansRcpp(x)
}
\arguments{
\item{x}{predictors matrix}
}
\value{
mean of x
}
\description{
Function to calculate column means of a NumericMatrix
}
