% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylm.R
\name{describe.data}
\alias{describe.data}
\title{Plot histograms of data and calculate common statistics}
\usage{
describe.data(data)
}
\arguments{
\item{data}{}
}
\value{
None
}
\description{
Plot histograms of data and calculate common statistics
}
\examples{
describe.data(data)
}
