% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_summary.R
\name{lm_summary}
\alias{lm_summary}
\title{lm_summary}
\usage{
lm_summary(model, Clevel = 0.95)
}
\arguments{
\item{model}{linear model object}

\item{Clevel}{the confidence level for the intervals. Default is 0.95.}
}
\value{
summary information of coefficients, residuals, and additional model information.
}
\description{
This function produces output similar to \code{summary()}, with an additional parameter for specifying the confidence level.
The result includes an extra column providing confidence intervals and an additional column indicating the significance of the p-value for each coefficient, determined by the specified confidence level.
}
\examples{
data(mtcars)
example_model <- lm(mpg ~ hp + wt, data = mtcars)

#obtain summary of the linear model
lm_summary(example_model)
}
