% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylm.R
\name{mylm.basic}
\alias{mylm.basic}
\title{the same function as the R function lm()}
\usage{
mylm.basic(
  formula,
  data,
  subset,
  weights,
  na.action,
  method = "qr",
  model = TRUE,
  x = FALSE,
  y = FALSE,
  qr = TRUE,
  singular.ok = TRUE,
  contrasts = NULL,
  offset,
  ...
)
}
\arguments{
\item{...}{}
}
\value{
lm object
}
\description{
the same function as the R function lm()
}
\examples{
mylm.basic(y~x1+x2,data)
}
