% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylm.R
\name{mylm.detailed}
\alias{mylm.detailed}
\title{Linear regression with summary and plots}
\usage{
mylm.detailed(formula, data)
}
\arguments{
\item{data}{}
}
\value{
lm model
}
\description{
Linear regression with summary and plots
}
\examples{
mylm.detailed(Y~x1+x2,data)
}
