% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylm.R
\name{mylm.predict}
\alias{mylm.predict}
\title{Make predictions and plot graphs using the incoming model}
\usage{
mylm.predict(model, x)
}
\arguments{
\item{x}{}
}
\value{
list
}
\description{
Make predictions and plot graphs using the incoming model
}
\examples{
mylm.predict(model, x)
}
