% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{obtain_f_stats}
\alias{obtain_f_stats}
\title{obtain_F_stats}
\usage{
obtain_f_stats(model)
}
\arguments{
\item{model}{linear model object}
}
\value{
F statistic, numerator degrees of freedom, residuals degrees of freedom, and the corresponding p-value.
}
\description{
This function gives F statistic, numerator degrees of freedom, residuals degrees of freedom, and the corresponding p-value of the linear model.
}
\examples{
data(mtcars)
example_model <- lm(mpg ~ hp + wt, data = mtcars)

#obtain F statistic and its p-value
obtain_f_stats(example_model)
}
