% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{obtain_p_value}
\alias{obtain_p_value}
\title{obtain_p_value}
\usage{
obtain_p_value(model)
}
\arguments{
\item{model}{linear model object}
}
\value{
p values of the linear model
}
\description{
obtain_p_value
}
