% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{obtain_r_square}
\alias{obtain_r_square}
\title{obtain_r_square}
\usage{
obtain_r_square(model)
}
\arguments{
\item{model}{linear model object}
}
\value{
R square and adjusted R square of the linear model
}
\description{
obtain_r_square
}
\examples{
#load the mtcars dataset
data(mtcars)

#lit a linear regression model
example_model <- lm(mpg ~ hp + wt, data = mtcars)

#obtain R square and adjusted R square
obtain_r_square(example_model)
}
