% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{obtain_residual_info}
\alias{obtain_residual_info}
\title{obtain_residual_info}
\usage{
obtain_residual_info(model)
}
\arguments{
\item{model}{linear model object}
}
\value{
residuals standard error and degrees of freedom of the linear model
}
\description{
This function provides the standard error and degrees of freedom for the residuals in the linear model.
}
\examples{
data(mtcars)
example_model <- lm(mpg ~ hp + wt, data = mtcars)

#obtain residual standard error and degrees of freedom
obtain_residual_info(example_model)
}
