% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{obtain_se}
\alias{obtain_se}
\title{obtain_se}
\usage{
obtain_se(model)
}
\arguments{
\item{model}{linear model object}
}
\value{
coefficients standard errors
}
\description{
This function provides the standard errors for each coefficient in the linear model.
}
\examples{
data(mtcars)
example_model <- lm(mpg ~ hp + wt, data = mtcars)

#obtain coefficients standard errors
obtain_se(example_model)
}
