% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{obtain_SSE}
\alias{obtain_SSE}
\title{obtain_SSE}
\usage{
obtain_SSE(model)
}
\arguments{
\item{model}{linear model object}
}
\value{
SSE of the linear model
}
\description{
This function gives the SSE of the linear model.
}
\examples{
data(mtcars)
example_model <- lm(mpg ~ hp + wt, data = mtcars)

#obtain SSE
obtain_SSE(example_model)
}
