% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{obtain_t_stats}
\alias{obtain_t_stats}
\title{obtain_t_stats}
\usage{
obtain_t_stats(model)
}
\arguments{
\item{model}{linear model object}
}
\value{
t statistics of the linear model
}
\description{
obtain_t_stats
}
\examples{
#load the mtcars dataset
data(mtcars)

#lit a linear regression model
example_model <- lm(mpg ~ hp + wt, data = mtcars)

#obtain t statistics
obtain_t_stats(example_model)
}
