% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_summary.R
\name{lm_summary}
\alias{lm_summary}
\title{lm_summary}
\usage{
lm_summary(model, Clevel = 0.95)
}
\arguments{
\item{model}{linear model object}

\item{Clevel}{the confidence level for the intervals. Default is 0.95.}
}
\value{
an summary table with information of coefficient and additional model information.
}
\description{
lm_summary
}
\examples{
#load the mtcars dataset
data(mtcars)

#fit a linear regression model
example_model <- lm(mpg ~ hp + wt, data = mtcars)

#obtain summary of the linear model
lm_summary(example_model)
}
