% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr.R
\name{lrm}
\alias{lrm}
\title{Multiple Linear Regression Model}
\usage{
lrm(y, x, Rcpp = F)
}
\arguments{
\item{y}{input value}

\item{x}{input value}

\item{Rcpp}{input value, "FALSE" by default}
}
\value{
the summary of linear regression model of x and y
}
\description{
Gets the mean of x, parameters, sse, t-value, p-value, significance, correlation, R square of linear regression model
When compare with the lm() function, our new function can have the same effect but speed up.
It is more efficient.
When there are NAN values, our model just simply remove the NAN value.
}
\examples{
n = 50
x1 = rnorm(n)
x2 = rnorm(n)
x3 = rnorm(n)
x = cbind(x1, x2, x3)
y = 4 + 3 * x1 + 2 * x2 - 1.5 * x3 + rnorm(n)
lrm(y,x)

}
