% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{obtain_f_stats}
\alias{obtain_f_stats}
\title{obtain_F_stats}
\usage{
obtain_f_stats(model)
}
\arguments{
\item{model}{linear model object}
}
\value{
F statistic and its p-value of the linear model
}
\description{
obtain_F_stats
}
