% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{obtain_residual_info}
\alias{obtain_residual_info}
\title{obtain_residual_info}
\usage{
obtain_residual_info(model)
}
\arguments{
\item{model}{linear model object}
}
\value{
residual standard error and degrees of freedom of the linear model
}
\description{
obtain_residual_info
}
