% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{obtain_se}
\alias{obtain_se}
\title{obtain_se}
\usage{
obtain_se(model)
}
\arguments{
\item{model}{linear model object}
}
\value{
coefficients standard errors of the linear model
}
\description{
obtain_se
}
